Change the location of modal entry to a new Hook
